Fix retarded c++ deduction

+++ CentoLib/include/Register.h
@@ -2,8 +2,8 @@
 #define REGISTER_H
 
 extern "C" {
-	void libcento_init(int hash);
-	void libcento_exit(int hash);
+	void libcento_init(size_t hash);
+	void libcento_exit(size_t hash);
 }
 
 #endif // REGISTER_H
