+++ CentoLib/src/Register.cpp
@@ -2,10 +2,10 @@
 #include "RuntimeObject.hpp"
 #include "CentoCommon.hpp"
 
-void libcento_init(int hash) {
+void libcento_init(size_t hash) {
 	CentoCommon::Singleton<RuntimeObject>::GetInstance()->LoadModule(hash);
 }
 
-void libcento_exit(int hash) {
+void libcento_exit(size_t hash) {
 	CentoCommon::Singleton<RuntimeObject>::GetInstance()->UnloadModule(hash);
 }
