+++ CentoLib/src/RuntimeObject.cpp
@@ -4,10 +4,10 @@ RuntimeObject::RuntimeObject() : _dll{ "CentoRT.dll" } {
 
 }
 
-void RuntimeObject::LoadModule(int hash) {
-	_dll.ExecFunction<void>("LoadModule", hash);
+void RuntimeObject::LoadModule(size_t hash) {
+	_dll.ExecFunction<void>("LoadModule", (int)hash);
 }
 
-void RuntimeObject::UnloadModule(int hash) {
-	_dll.ExecFunction<void>("UnloadModule", hash);
+void RuntimeObject::UnloadModule(size_t hash) {
+	_dll.ExecFunction<void>("UnloadModule", (int)hash);
 }
