+++ CentoRT/src/Exports.cpp
@@ -11,10 +11,10 @@ CENTO_API void Shutdown()
     std::cout << "[CentoRT] Shutdown.\n";
 }
 
-CENTO_API void LoadModule(int hash) {
+CENTO_API void LoadModule(size_t hash) {
     std::cout << "[CentoRT] Module loaded: " << hash << "\n";
 }
 
-CENTO_API void UnloadModule(int hash) {
+CENTO_API void UnloadModule(size_t hash) {
     std::cout << "[CentoRT] Module unloaded: " << hash << "\n";
 }
